refactoring into a ruledb package
